Extracts FilterSlider components and uses CSS Modules and tokens
