+++ src/components/search/Filter/Facets.tsx
@@ -1,12 +1,14 @@
 import { Label as UILabel, List as UIList } from '@faststore/ui'
 
+import type {
+  Filter_FacetsFragment,
+  IStoreSelectedFacet,
+} from '@generated/graphql'
 import Accordion, { AccordionItem } from 'src/components/ui/Accordion'
 import { Badge } from 'src/components/ui/Badge'
 import Checkbox from 'src/components/ui/Checkbox'
-import type {
-  IStoreSelectedFacet,
-  Filter_FacetsFragment,
-} from '@generated/graphql'
+
+import styles from './facets.module.scss'
 
 interface FacetsProps {
   testId: string
@@ -24,8 +26,10 @@ function Facets({
   onAccordionChange,
 }: FacetsProps) {
   return (
-    <div className="filter" data-store-filter data-testid={testId}>
-      <h2 className="text__title-mini-alt">Filters</h2>
+    <div className={styles.fsFacets} data-testid={testId}>
+      <h2 className="text__title-mini-alt" data-fs-facets-title>
+        Filters
+      </h2>
       <Accordion expandedIndices={indicesExpanded} onChange={onAccordionChange}>
         {facets.map(({ label, values, key }, index) => (
           <AccordionItem
@@ -40,17 +44,23 @@ function Facets({
                 const id = `${testId}-${label}-${item.label}`
 
                 return (
-                  <li key={id} className="filter__item">
+                  <li key={id} data-fs-facet-item>
                     <Checkbox
                       id={id}
                       checked={item.selected}
                       onChange={() => onFacetChange({ key, value: item.value })}
+                      data-fs-facet-item-checkbox
                       data-testid={`${testId}-accordion-panel-checkbox`}
                       data-value={item.value}
                       data-quantity={item.quantity}
                     />
-                    <UILabel htmlFor={id} className="text__title-mini-alt">
-                      {item.label} <Badge>{item.quantity}</Badge>
+                    <UILabel
+                      htmlFor={id}
+                      className="text__title-mini-alt"
+                      data-fs-facet-item-label
+                    >
+                      {item.label}{' '}
+                      <Badge data-fs-facet-item-badge>{item.quantity}</Badge>
                     </UILabel>
                   </li>
                 )
