+++ src/components/search/Filter/index.ts
@@ -1,1 +1,2 @@
 export { default } from './Filter'
+export { default as FilterSlider } from './FilterSlider'
