+++ src/styles/global/components.scss
@@ -12,7 +12,6 @@
 @import "src/components/regionalization/RegionalizationModal/regionalization-modal.scss";
 
 // Search
-@import "src/components/search/Filter/filter.scss";
 @import "src/components/search/Suggestions/suggestions.scss";
 @import "src/components/search/SuggestionProductCard/suggestion-product-card.scss";
