+++ src/styles/global/storybook-components.scss
@@ -9,7 +9,6 @@
 @import "src/components/common/Navbar/navbar.scss";
 
 // Search
-@import "src/components/search/Filter/filter.scss";
 @import "src/components/search/Suggestions/suggestions.scss";
 @import "src/components/search/SuggestionProductCard/suggestion-product-card.scss";
